haskell-hoogle: Update pre-shipped keyword.txt file
Previously, 'keyword.txt' was a simple dump of the
http://www.haskell.org/haskellwiki/Keywords page, and hoogle was converting it
(in update-hoogle) to the .txt/.hoo format. The new method which doesn't use
'data' anymore needs a pre-generated '.txt' file, which hoogle is able to
convert to .hoo.

This file has been taken from a "full" (i.e. using "hoogle data") database, so
its contents is frozen both in time (2014-09-21) and in the version of hoogle
used (4.2.33).