haskell-hoogle: Rework how the database is generated
This finally updates the update-hoogle script to the new way of handling
database in newer hoogle. Unfortunately, we do this by completely skipping the
'data' subcommand, since that wants a fixed list of (big) files, and we instead
convert and combine the packages manually. The script did this before, but only
for the local database; the 'keyword' db was handled by 'data', which is now
also moved to convert/combine.

This patch needs an updated keyword.txt file.