gitit: Use dh_link instead of dh_linktree
dh_linktree imposes a strong dependency against libjs-jquery. This
means that every time a new version of libjs-jquery gets uploaded,
libghc-gitit-data will have broken dependencies.

This seems to be a known problem:
    https://bugs.debian.org/676455
    https://bugs.debian.org/784859
    https://bugs.debian.org/820879

We already replace the jquery shipped by gitit with the much newer
version available to the archive without validating if they are
compatible, so there is no reason to impose strict dependencies against
newer versions of libjs-jquery.

Use dh_link instead of dh_linktree to create the symlinks and avoid
continuous uploading of the package:
    https://bugs.debian.org/796809