ghc: drop wrong code reducing parallelism
Drop the export DEB_BUILD_OPTIONS, this is not the correct way to reduce
parallelism. It prevents users to use a lower level of build parallelism
(for example 1 in that case). The dh --max-parallel=maximum option
should be used instead.

Anyway in that case reducing the parallelism doesn't bring anything, it
just make the build slower to succeed or fail. Therefore just drop the
code.