ghc: Re-generate the stage2 package cache
I have noticed that on some rare cases the package cache generated for
stage2 is stale. This is probably an upstream bug, and it wouldn't
affect us (the package cache for binary-dist-dir is correct), but we use
ghc-pkg from stage2 when we generate our control file. Maybe we should
consider using ghc-pkg from binary-dist-dir instead.

As a work-around for now, regenerate the stage2 package cache.

Also, drop the hadrian-no-pkg-recache patch for now, since we don't
support cross building GHC yet.